+++ src/export/ExportMP2.cpp
@@ -276,8 +276,8 @@ int ExportMP2::Export(AudacityProject *project,
 
    ProgressDialog *progress = new ProgressDialog(wxFileName(fName).GetName(),
       selectionOnly ?
-      wxString::Format(_("Exporting selected audio at %d kbps"), bitrate) :
-      wxString::Format(_("Exporting entire file at %d kbps"), bitrate));
+      wxString::Format(_("Exporting selected audio at %d kbps"), (int) bitrate) :
+      wxString::Format(_("Exporting entire file at %d kbps"), (int) bitrate));
 
    int updateResult = eProgressSuccess;
    while(updateResult == eProgressSuccess) {
